requirements.txt
pandas
tensorflow
flask
numpy
keras
Pillow
flask-cors  